chore(apps/legacy-api): move supported networks into NetworkValidationPipe
