--- apps/legacy-api/src/common/networks.ts
@@ -1,1 +1,0 @@
-export type SupportedNetwork = 'mainnet' | 'testnet'
