+++ apps/legacy-api/src/controllers/PoolPairController.ts
@@ -1,8 +1,7 @@
 import { Controller, Get, Query } from '@nestjs/common'
 import { PoolPairData } from '@defichain/whale-api-client/dist/api/poolpairs'
 import { WhaleApiClientProvider } from '../providers/WhaleApiClientProvider'
-import { NetworkValidationPipe } from '../pipes/NetworkValidationPipe'
-import { SupportedNetwork } from '../common/networks'
+import { NetworkValidationPipe, SupportedNetwork } from '../pipes/NetworkValidationPipe'
 
 @Controller('v1')
 export class PoolPairController {
